add 添加一些用户常量 add JSON 工具类新增方法 update 将原来的根据用户昵称查询用户(username 可以不传)拆分为: 根据用户昵称查询用户(username 为必传参数)、查询所有用户 update 修改 Tag, 权值由浮点数改为整数 update 修改根据标签搜索用户, 因为 user 的 tag:字段需要额外存储权值字段

+++ src/main/java/com/guyi/kindredspirits/model/domain/Tag.java
@@ -38,12 +38,12 @@ public class Tag implements Serializable {
     /**
      * 基本权值
      */
-    private Double baseWeight;
+    private Integer baseWeight;
 
     /**
      * 父标签权值+自己的基本权值
      */
-    private Double weights;
+    private Integer weights;
 
     /**
      * 创建时间
