+++ src/main/java/com/guyi/kindredspirits/model/request/TagAddRequest.java
@@ -32,11 +32,11 @@ public class TagAddRequest implements Serializable {
     /**
      * 基本权值
      */
-    private Double baseWeight;
+    private Integer baseWeight;
 
     /**
      * 父标签权值+自己的基本权值
      */
-    private Double weights;
+    private Integer weights;
 
 }
